adds email file
